moves TabularData from musana.loader to dimcat.dtypes
